requirements.txt
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
defusedxml==0.7.1
Django==5.2.8
django-allauth==65.13.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
httplib2==0.31.0
idna==3.11
oauthlib==3.3.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.2.5
python-decouple==3.8
python-dotenv==1.2.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.4
tzdata==2025.2
urllib3==2.5.0
